Refactor: 스타일 분리
