+++ src/components/InputController/styles.ts
@@ -1,0 +1,67 @@
+import styled from "styled-components";
+import {Flexbox} from "@/styles/util";
+
+export const InputControllerLayout = styled.div`
+  ${Flexbox};
+  width: 99%;
+`;
+
+export const InputLayer = styled.div`
+  ${Flexbox};
+  font-size: ${({ theme }) => theme.fontSize.xl};
+  height: 30px;
+`;
+
+export const InputFormLayout = styled.form`
+  width: 100%;
+`;
+
+export const Input = styled.input.attrs({ type: 'number', placeholder: '0' })`
+  font-size: ${({ theme }) => theme.fontSize.xl};
+  width: 100%;
+  height: 100%;
+  text-align: right;
+  padding: 0 5px;
+  border-bottom: 1px solid ${({ theme }) => theme.color.black};
+
+  &::-webkit-outer-spin-button,
+  &::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+  -moz-appearance: textfield;
+`;
+
+export const InputAmount = styled.div`
+  font-size: ${({ theme }) => theme.fontSize.xl};
+  width: 100%;
+  text-align: right;
+  padding: 4px;
+  border: 1px solid ${({ theme }) => theme.color.black};
+  cursor: pointer;
+  transition: background-color 200ms;
+
+  &:hover {
+    background-color: ${({ theme }) => theme.color.darkBlack};
+  }
+
+  &:active {
+    background-color: ${({ theme }) => theme.color.lightBlack};
+  }
+`;
+
+export const ReturnButton = styled.button.attrs({ type: 'button' })`
+  font-size: ${({ theme }) => theme.fontSize.xl};
+  padding: 5px;
+  border: 1px solid ${({ theme }) => theme.color.black};
+  cursor: pointer;
+  transition: background-color 200ms;
+
+  &:hover {
+    background-color: ${({ theme }) => theme.color.darkBlack};
+  }
+
+  &:active {
+    background-color: ${({ theme }) => theme.color.lightBlack};
+  }
+`;
